Moved functions under utilities along with the restart cancel function.
